backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.3
SQLAlchemy==2.0.31
python-dotenv==1.0.1
gunicorn==22.0.0
requests==2.32.5
